Type returned by sensor_ReadByte switches from byte to int
git-svn-id: http://panstamp.googlecode.com/svn/trunk@716 d77f469f-0316-7e13-d0a9-da7f422f7f4c

+++ arduino/sketches/temphumpress/sensor.ino
@@ -48,7 +48,7 @@ byte sensor_ReadByte(void);
  *
  * Read data byte from DHT11 sensor
  */
-byte sensor_ReadByte(void)
+int sensor_ReadByte(void)
 {
   byte i, result = 0;
   unsigned int count = 20000;
